+++ Base/FUI/Panels/Datasets/datasetsPanel.hpp
@@ -3,8 +3,12 @@
 
 #include <QWidget>
 #include <QGroupBox>
+#include <QFrame>
 
-#include "datasetsLayout.hpp"
+#include "datasetsToolbarLayout.hpp"
+#include "datasetsTreeLayout.hpp"
+
+#include <CustomWidgets/fluoLine.hpp>
 
 class DatasetsPanel : public QWidget
 {
@@ -13,15 +17,29 @@ class DatasetsPanel : public QWidget
   public:
     DatasetsPanel()
     {
-      datasetsFrame->setLayout(datasetsLayout);
-      frameLayout->addWidget(datasetsFrame);
-      this->setLayout(frameLayout);
+      topFrame->setLayout(toolLayout);
+      botFrame->setLayout(treeLayout);
+      frameLayout->addWidget(topFrame,0,0);
+      frameLayout->addWidget(sepLine,1,0);
+      frameLayout->addWidget(botFrame,2,0);
+      datasetsFrame->setLayout(frameLayout);
+      widgetLayout->addWidget(datasetsFrame);
+      this->setLayout(widgetLayout);
     }
 
   private:
     QGroupBox *datasetsFrame = new QGroupBox();
+
+    QFrame *topFrame = new QFrame();
+    QFrame *botFrame = new QFrame();
+
+    DatasetsToolbarLayout *toolLayout = new DatasetsToolbarLayout();
+    DatasetsTreeLayout *treeLayout = new DatasetsTreeLayout();
+
+    QGridLayout *widgetLayout = new QGridLayout();
     QGridLayout *frameLayout = new QGridLayout();
-    DatasetsLayout *datasetsLayout = new DatasetsLayout();
+
+    FluoLine *sepLine = new FluoLine();
 };
 
 #endif
